save_data_to_db_in_second_fragment

+++ app/src/main/java/ru/easygraphics/chartsettingsWindow/ChartDescriptionScreen.kt
@@ -4,6 +4,6 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentFactory
 import com.github.terrakok.cicerone.androidx.FragmentScreen
 
-class ChartDescriptionScreen(val chart_id:Int): FragmentScreen {
+class ChartDescriptionScreen(val chart_id:Long): FragmentScreen {
     override fun createFragment(factory: FragmentFactory): Fragment = ChartDescriptionFragment.newInstance(chart_id)
 }
